consolidate tbc into tbc.go
